+++ tests/testthat/test-tax_range_space.R
@@ -1,4 +1,4 @@
-test_that("tax_range_soace() works", {
+test_that("tax_range_space() works", {
   # Grab internal data and set-up
   occdf <- tetrapods
   # n unique taxa
@@ -17,21 +17,19 @@ test_that("tax_range_soace() works", {
     nrow(tax_range_space(occdf = occdf, name = "accepted_name",
                          method = "occ")),
     unique_taxa)
-
-  # Expect greater than
-  expect_gt(nrow(tax_range_space(occdf = occdf,
+  expect_equal(nrow(tax_range_space(occdf = occdf,
                                name = "accepted_name",
                                method = "gcd")),
                 unique_taxa)
-  expect_gt(nrow(tax_range_space(occdf = occdf,
+  expect_equal(nrow(tax_range_space(occdf = occdf,
                                name = "accepted_name",
                                method = "gcd")),
             unique_taxa)
-  expect_gt(nrow(tax_range_space(occdf = occdf,
+  expect_equal(nrow(tax_range_space(occdf = occdf,
                                name = "accepted_name",
                                method = "con")),
             unique_taxa)
-  expect_gt(nrow(tax_range_space(occdf = occdf, name = "accepted_name",
+  expect_equal(nrow(tax_range_space(occdf = occdf, name = "accepted_name",
                                method = "con")),
             unique_taxa)