+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * palaeoverse now requires deeptime (>= 1.0.0)
 * Added additional test to catch when Zenodo is offline (palaeorotate)
+* Added a coords argument to the tax_range_space function to control whether coordinates are returned
 * palaeorotate documentation updated to reflect updates to GPlates Web Service (#74)
 * MULLER2022 plate rotation model made available in palaeorotate (#74)
 * model argument in palaeorotate now allows multiple models to be called - palaeogeographic uncertainty is now calculated on the requested models